Finish projectile bounce and auto destroy

+++ The Tank Arena/Src/Game/Components/Sprite/LevelSpriteComponent.hpp
@@ -12,5 +12,5 @@ struct LevelSpriteComponent : lic::Component
 {
 	std::vector<sf::Sprite> tile_sprites;
 
-	LevelSpriteComponent(const ProgramInfo& program_info, const lio::Matrix<Cell>& level, const std::string& tileset_id, int tile_size);
+	LevelSpriteComponent(const ProgramInfo& program_info, const lio::Matrix<Cell>& level, const std::string& tileset_id, size_t tile_size);
 };
